requirements.txt
# IA et génération d'exercices
ollama==0.4.4

# API REST
flask==3.0.0
flask-cors==4.0.0

# Tests (optionnel)
pytest==7.4.3
pytest-cov==4.1.0

# Gestion de configuration
python-dotenv==1.0.0
